Extract ephemeral machine code from `fly console`
The intent is to make it reusable.
